CJAL-24/09/2018 - termina fase de desarrollo Modulo Clientes

+++ 1.Layers/1.3Application/1.3.2MainModule/Application.MainModule/Servicios/AccesoADatos/EstadoRDataAccess.cs
@@ -22,7 +22,7 @@ namespace Application.MainModule.Servicios.AccesoADatos
             return uow.Repository<EstadosRepublica>().GetAll().ToList();
         }
 
-        public EstadosRepublica BuscarIdEdo(int idEdo)
+        public EstadosRepublica BuscarIdEdo(byte idEdo)
         {
             return uow.Repository<EstadosRepublica>().GetSingle(x => x.IdEstadoRep.Equals(idEdo));
         }
