+++ 1.Layers/1.3Application/1.3.2MainModule/Application.MainModule/Servicios/AccesoADatos/PaisDataAccess.cs
@@ -22,7 +22,7 @@ namespace Application.MainModule.Servicios.AccesoADatos
             return uow.Repository<Pais>().GetAll().ToList();
         }
 
-        public Pais BuscarIdPais(int idPais)
+        public Pais BuscarIdPais(byte idPais)
         {
             return uow.Repository<Pais>().GetSingle(x => x.IdPais.Equals(idPais));
         }
